refactor: rename addChangeOutputToPsbt to addWalletOutputToPsbt, move
It is only accurate to call it "change output" if the method verifies
all the inputs, otherwise we can only say that it is an output that
belongs to the wallet passed as an argument.

Move method from WalletUnspent.ts to WalletOutput.ts.

Issue: BG-47824
BREAKING CHANGE: addChangeOutputToPsbt renamed to addWalletOutputToPsbt
