+++ modules/utxo-lib/src/bitgo/wallet/index.ts
@@ -1,5 +1,6 @@
 export * from './chains';
 export * from './Unspent';
+export * from './WalletOutput';
 export * from './WalletUnspentSigner';
 export * from './WalletScripts';
 export * from './WalletKeys';
