+++ modules/utxo-lib/test/bitgo/wallet/WalletUnspent.ts
@@ -26,8 +26,8 @@ import {
   createPsbtForNetwork,
   createPsbtFromTransaction,
   addToPsbt,
+  addWalletOutputToPsbt,
   toPrevOutput,
-  addChangeOutputToPsbt,
   KeyName,
 } from '../../../src/bitgo';
 
@@ -136,8 +136,7 @@ describe('WalletUnspent', function () {
   ): Transaction<bigint> {
     const psbt = createPsbtForNetwork({ network });
     const total = BigInt(unspentSum<bigint>(unspents, 'bigint'));
-    // Kinda weird, treating entire value as change, but tests the relevant paths
-    addChangeOutputToPsbt(psbt, walletKeys, getInternalChainCode(scriptType), CHANGE_INDEX, total - FEE);
+    addWalletOutputToPsbt(psbt, walletKeys, getInternalChainCode(scriptType), CHANGE_INDEX, total - FEE);
 
     unspents.forEach((u) => {
       addToPsbt(psbt, u, walletKeys, signer, cosigner, network);
